requirements.txt
fastapi
uvicorn
sentencepiece
transformers
torch 
torchvision 
torchaudio
httpie
langchain==0.3.7
langchain-community==0.3.5
openai==0.28.0
python-dotenv
